horoscope/tg-horoscope/src/requirements.txt
databases==0.8.0
fastapi==0.104.0
pydantic==2.4.2
python-telegram-bot==20.6
pydantic_core==2.10.1
Requests==2.31.0
sentry_sdk==1.32.0
SQLAlchemy==1.4.49
uvicorn==0.23.2
asyncpg==0.28.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
alembic==1.12.1
python-dotenv
python-telegram-bot[rate-limiter]
